Edited according to @dustin-goldenowl's review
