+++ lib/l10n/app_en.arb
@@ -88,5 +88,6 @@
     "save_job_page_share": "Share",
     "save_job_page_send_message": "Send message",
     "save_job_page_delete_all": "Delete all",
+    "save_job_page_feature_not_yet_released": "Feature not yet released",
     "save_job_page_delete": "Delete"
 }
