+++ lib/l10n/app_vi.arb
@@ -88,5 +88,6 @@
     "save_job_page_share": "Chia sẻ",
     "save_job_page_send_message": "Gửi tin nhắn",
     "save_job_page_delete_all": "Xóa tất cả",
+    "save_job_page_feature_not_yet_released": "Tính năng chưa ra mắt",
     "save_job_page_delete": "Xóa"
 }
