+++ lib/src/presentations/save_job/cubit/save_job_state.dart
@@ -1,0 +1,8 @@
+part of 'save_job_cubit.dart';
+
+class SaveJobState extends Equatable {
+  const SaveJobState();
+
+  @override
+  List<Object> get props => [];
+}
